+++ CopilotKit/packages/react-core/CHANGELOG.md
@@ -1,5 +1,32 @@
 # ui
 
+## 1.4.1-pre.2
+
+### Patch Changes
+
+- Add convertActionsToDynamicStructuredTools to sdk-js
+- Updated dependencies
+  - @copilotkit/runtime-client-gql@1.4.1-pre.2
+  - @copilotkit/shared@1.4.1-pre.2
+
+## 1.4.1-pre.1
+
+### Patch Changes
+
+- add zod conversion
+- Updated dependencies
+  - @copilotkit/runtime-client-gql@1.4.1-pre.1
+  - @copilotkit/shared@1.4.1-pre.1
+
+## 1.4.1-pre.0
+
+### Patch Changes
+
+- lower case copilotkit property
+- Updated dependencies
+  - @copilotkit/runtime-client-gql@1.4.1-pre.0
+  - @copilotkit/shared@1.4.1-pre.0
+
 ## 1.4.0
 
 ### Minor Changes